+++ src/card/cards/stale_followups.rs
@@ -41,3 +41,5 @@ pub const SPIRIT_SHIELD: CardDefinitionId = CardDefinitionId(1666);
 pub const ZELYON_SWORD: CardDefinitionId = CardDefinitionId(1667);
 pub const GIANT_TORTOISE: CardDefinitionId = CardDefinitionId(1668);
 pub const ELVISH_HUNTER: CardDefinitionId = CardDefinitionId(1669);
+pub const ENCHANTED_BEING: CardDefinitionId = CardDefinitionId(1670);
+pub const DEMONIC_TORMENT: CardDefinitionId = CardDefinitionId(1671);