+++ src/game/ability_layers.rs
@@ -293,7 +293,9 @@ impl Game {
             | AppliedEffectDef::CannotBeBlockedBy(_)
             | AppliedEffectDef::CanBlockOnly(_)
             | AppliedEffectDef::PreventDamageFrom(_)
+            | AppliedEffectDef::PreventCombatDamageFrom(_)
             | AppliedEffectDef::PreventCombatDamage
+            | AppliedEffectDef::PreventCombatDamageDealtBy
             | AppliedEffectDef::AddLandTypes(_)
             | AppliedEffectDef::SetLandTypes(_)
             | AppliedEffectDef::Animate(_)