+++ src/game/entry_replacements.rs
@@ -588,7 +588,9 @@ impl Game {
             | AppliedEffectDef::CannotBeBlockedBy(_)
             | AppliedEffectDef::CanBlockOnly(_)
             | AppliedEffectDef::PreventDamageFrom(_)
+            | AppliedEffectDef::PreventCombatDamageFrom(_)
             | AppliedEffectDef::PreventCombatDamage
+            | AppliedEffectDef::PreventCombatDamageDealtBy
             | AppliedEffectDef::AddLandTypes(_)
             | AppliedEffectDef::SetLandTypes(_)
             | AppliedEffectDef::RemoveAbilities(_)